Split test file into multiple files
one for each function in optimizedForceMove.sol
