+++ test/optimizedForceMove/_recoverSigner.test.ts
@@ -1,0 +1,54 @@
+import {ethers} from 'ethers';
+import {expectRevert} from 'magmo-devtools';
+// @ts-ignore
+import optimizedForceMoveArtifact from '../../build/contracts/TESTOptimizedForceMove.json';
+import {splitSignature, arrayify} from 'ethers/utils';
+
+let networkId;
+let optimizedForceMove: ethers.Contract;
+const provider = new ethers.providers.JsonRpcProvider(
+  `http://localhost:${process.env.DEV_GANACHE_PORT}`,
+);
+const signer = provider.getSigner(0);
+async function setupContracts() {
+  networkId = (await provider.getNetwork()).chainId;
+  const optimizedForceMoveContractAddress = optimizedForceMoveArtifact.networks[networkId].address;
+  optimizedForceMove = new ethers.Contract(
+    optimizedForceMoveContractAddress,
+    optimizedForceMoveArtifact.abi,
+    signer,
+  );
+}
+
+const participants = ['', '', ''];
+const wallets = new Array(3);
+
+// populate wallets and participants array
+for (let i = 0; i < 3; i++) {
+  wallets[i] = ethers.Wallet.createRandom();
+  participants[i] = wallets[i].address;
+}
+
+beforeAll(async () => {
+  await setupContracts();
+});
+
+async function sign(wallet: ethers.Wallet, msgHash: string | Uint8Array) {
+  // msgHash is a hex string
+  // returns an object with v, r, and s properties.
+  return splitSignature(await wallet.signMessage(arrayify(msgHash)));
+}
+
+describe('_recoverSigner', () => {
+  it('recovers the signer correctly', async () => {
+    // following https://docs.ethers.io/ethers.js/html/cookbook-signing.html
+    const privateKey = '0x0123456789012345678901234567890123456789012345678901234567890123';
+    const wallet = new ethers.Wallet(privateKey);
+    const msgHash = ethers.utils.id('Hello World');
+    const msgHashBytes = arrayify(msgHash);
+    const sig = await sign(wallet, msgHashBytes);
+    expect(await optimizedForceMove.recoverSigner(msgHash, sig.v, sig.r, sig.s)).toEqual(
+      wallet.address,
+    );
+  });
+});
